pyproject.toml
[build-system]
requires      = ["flit_core == 3.9.0"]
build-backend =  "flit_core.buildapi"

[project]
name        = "powerviz"
version     = "0.1.0"
description = "Simple dashboard showing ISO/RTO and EIA data."

authors = [
  { name = "Matthew Pham", email = "mtpham.github@quantamail.net" },
]

readme = "README.md"
license = {file = "LICENSE.md"}

requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "pandas[excel] >= 2.2.0",  # calamine engine for excel (2.2.0)
    "pytz",
    "tqdm",
    "aiohttp",
    "tenacity",
    "psycopg2-binary",
    "dash",
    "dash-bootstrap-components",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "mypy",
    "pre-commit",
    "pylint"
]

[project.urls]
"Homepage"    = "https://github.com/mtpham99/powerviz"
"Source"      = "https://github.com/mtpham99/powerviz"
"Bug Tracker" = "https://github.com/mtpham99/powerviz/issues"

[tool.black]
line-length = 79
include = ".pyi?$"
exclude = """
/(
    .git
    | .hg
    | .mypy_cache
    | .tox
    | .venv
    | _build
    | buck-out
    | build
    | dist
)/
"""

[tool.isort]
line_length = 79
multi_line_output = 3
include_trailing_comma = true

[tool.mypy]
disallow_untyped_defs = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.pylint]
max-line-length = 79
disable = [
    "fixme",
    "missing-class-docstring",
    "missing-function-docstring",
    "missing-module-docstring",
    "too-few-public-methods",
    "too-many-arguments",
    "too-many-function-args",
    "too-many-locals",
    "unnecessary-pass"
]
